# mcp-erp-demo/sales-orders-java-rest-crud-postgres/src/main/java/com/example/salesorder/exception/ErrorResponse.java
package com.example.salesorder.exception;

import com.fasterxml.jackson.annotation.JsonInclude;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.Map;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
@Schema(description = "Standard error response")
public class ErrorResponse {

    @Schema(description = "Timestamp when the error occurred", example = "2025-03-11T09:22:00")
    private LocalDateTime timestamp;

    @Schema(description = "HTTP status code", example = "404")
    private int status;

    @Schema(description = "Error type", example = "Not Found")
    private String error;

    @Schema(description = "Error message", example = "Order not found with id: 1")
    private String message;

    @Schema(description = "Detailed error information for validation errors")
    private Map<String, String> details;
}